+++ src/eca/messenger.clj
@@ -2,5 +2,7 @@
   "An interface for sending messages to a 'client',
    whether it's a editor or a no-op for test.")
 
+(set! *warn-on-reflection* true)
+
 (defprotocol IMessenger
   (chat-content-received [this data]))