+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- Fix ollama servers discovery
+
 ## 0.0.2
 
 - First alpha release